Fixed crash with unloading window
If both power saving options are off, the watchface will crash when unloading,
because it tries to unsubscribe from the health service events, but it never
subscribed in the first place!

Also fixed unsubscribing from the tick timer service, since the wrong parameter was passed to unsubscribe when unloading the window, and it wasn't protected by a NULL pointer check (it shouldn't be NULL, but just in case, it's better to check).