Removed DisplaySeconds added units instead
Most of the time when DisplaySeconds is checked, it's to determine which
units to use for the tick timer.  So we should be caching that
information, rather than litter the code with a bunch of if statements.
The most frequent check of DisplaySeconds was really to determine if
seconds should be displayed or not, but the replacement condition isn't
any worse in terms of performance.